Shader - Still not complete.
I have the Shader isolated in its own folder and in a test scene as to not ruin anything, its still a bit tempermental. I honestly have no clue whats happening with the shader code, I was trying out random shit really and well This is the result. I will try and give a rundown on how it works as I did not make any good coments.